+++ go-corona-app/buttons/Next.js
@@ -1,5 +1,6 @@
 import React from "react";
-import { Button } from "react-native";
+import { Button } from 'react-native-material-ui'
+import { StyleSheet, Text, View, Image, ScrollView } from 'react-native';
 
 class Next extends React.Component {
   render() {
@@ -7,10 +8,7 @@ class Next extends React.Component {
     if (isActive === false) return null;
 
     return (
-        <Button
-        title="Next"
-        onPress={()  => this.props.goToNextStep()}
-      />
+        <Button raised primary onPress={() => this.props.goToNextStep()} text="Next" style={{container: {width:100,backgroundColor:'#E03D51', marginBottom:'2%', paddingBottom:'10%',  }}}/>
     );
   }
 }
