+++ go-corona-app/buttons/Submit.js
@@ -1,5 +1,7 @@
 import React from "react";
-import { Button, Alert } from "react-native";
+import { Alert } from "react-native";
+import { Button } from 'react-native-material-ui'
+
 
 class Submit extends React.Component {
 
@@ -8,10 +10,10 @@ class Submit extends React.Component {
     if (isActive === false) return null;
 
     return (
-
-        <Button
-        title="Submit"
+        <Button raised primary
+        text="Submit"
         onPress={()  => Alert.alert('Submit Button pressed')}
+        style={{container: {width:100,backgroundColor:'#E03D51', marginBottom:'2%'}}}
       />
     );
