+++ go-corona-app/screens/checkup/Introduction.js
@@ -4,24 +4,50 @@ import * as React from 'react';
 import { StyleSheet, Text, View, Image, ScrollView } from 'react-native';
 import { RectButton } from 'react-native-gesture-handler';
 import CheckupIntroductionImage from '../../assets/images/CheckupIntroduction.svg'
+import  { StepList } from '../../stepper/StepList';
+import { Step }  from '../../stepper/Step';
+import CheckupTerms from './Terms';
 
+const PageOne = () => (
+  <ScrollView>
+  <View>
+    <CheckupIntroductionImage style={styles.image} width="200" height="200" />
+    <View style={styles.introduction}>
+      <Text style={styles.title}>Hello!</Text>
+      <Text style={styles.description}>
+        You are about to use a short (3min), safe and anonymous health checkup.
+        Your answers will be carefully analyzed and you will learn about possible causes of your symptoms.
+        You are adviced not to rush to the hospital if you have mild symptoms.
+        This is only to help you self-quarantine/isolate yourself from others based on how long you have had these symptoms for.
+      </Text>
+    </View>
+  </View>
+</ScrollView>
+   
+ )
+ 
+ const PageTwo = () => (
+   <View><Text>We have a second page.</Text></View>
+    
+  )
+ 
+  const PageThree = () => (
+   <View><Text>We have a third page.</Text></View>
+    
+  )
 
 export default function CheckupIntroduction() {
   return (
-    <ScrollView>
+   
       <View style={styles.container}>
-        <CheckupIntroductionImage style={styles.image} width="200" height="200" />
-        <View style={styles.introduction}>
-          <Text style={styles.title}>Hello!</Text>
-          <Text style={styles.description}>
-            You are about to use a short (3min), safe and anonymous health checkup.
-            Your answers will be carefully analyzed and you will learn about possible causes of your symptoms.
-            You are adviced not to rush to the hospital if you have mild symptoms.
-            This is only to help you self-quarantine/isolate yourself from others based on how long you have had these symptoms for.
-          </Text>
+          <StepList>
+          <Step component={PageOne} />
+          <Step component={CheckupTerms} />
+          <Step component={PageThree} />
+          </StepList>
         </View>
-      </View>
-    </ScrollView>
+      
+    
   );
 }
